safeplate-backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
passlib[bcrypt]
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
boto3>=1.26.0
python-multipart>=0.0.6
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
python-dotenv
boto3>=1.26.0
email-validator>=2.0.0
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
python-dotenv
boto3>=1.26.0
email-validator>=2.0.0
httpx>=0.24.0
cachetools>=5.3.0